Removed useless(?) event listener
This event's mixin redirects relevant usages of World.isDay(). isDay() is calculated based on whether the dimension has fixed time, and the world's ambientDarkness value, which is also influenced be the weather. (This is why you can sleep during thunderstorms, and can go to sleep a whole 26.6 seconds earlier in rainy weather.)
The latest daytime vanillaResult could be false at is 12541 (in clear weather), so it seems like there won't ever be a case in vanilla world types where the if statement is true.
Adding this here seems like it just causes issues with other mods that change the length of the day, or add dimensions whit fixed time, where beds don't explode.